+++ hutool-extra/src/main/java/cn/hutool/extra/template/engine/velocity/VelocityTemplate.java
@@ -8,7 +8,9 @@ import java.util.Map;
 import org.apache.velocity.VelocityContext;
 import org.apache.velocity.app.Velocity;
 
+import cn.hutool.core.convert.Convert;
 import cn.hutool.core.io.IoUtil;
+import cn.hutool.core.lang.TypeReference;
 import cn.hutool.core.util.CharsetUtil;
 import cn.hutool.core.util.StrUtil;
 import cn.hutool.extra.template.AbstractTemplate;
@@ -65,17 +67,15 @@ public class VelocityTemplate extends AbstractTemplate implements Serializable {
 	 * @return {@link VelocityContext}
 	 */
 	private VelocityContext toContext(Map<?, ?> bindingMap) {
-		return new VelocityContext(bindingMap);
+		final Map<String, Object> map = Convert.convert(new TypeReference<Map<String, Object>>() {}, bindingMap);
+		return new VelocityContext(map);
 	}
 	
 	/**
 	 * 加载可用的Velocity中预定义的编码
 	 */
 	private void loadEncoding() {
-		String charset = (String) Velocity.getProperty(Velocity.OUTPUT_ENCODING);
-		if(StrUtil.isEmpty(charset)) {
-			charset = (String) Velocity.getProperty(Velocity.INPUT_ENCODING);
-		}
+		final String charset = (String) Velocity.getProperty(Velocity.INPUT_ENCODING);
 		this.charset = StrUtil.isEmpty(charset) ? CharsetUtil.UTF_8 : charset;
 	}
 }